+++ src/GenerateProgrammaticPipelineFromConfig/Command.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-tooling for the canonical source repository
- * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2016-2017 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-tooling/blob/master/LICENSE.md New BSD License
  */
 
@@ -67,7 +67,7 @@ class Command
         );
 
         try {
-            $generator = new Generator();
+            $generator = new Generator($this->console);
             $generator->projectDir = $this->projectDir;
             $generator->process($this->locateConfigFile($args));
         } catch (GeneratorException $e) {