+++ src/GenerateProgrammaticPipelineFromConfig/Generator.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-tooling for the canonical source repository
- * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2016-2017 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-tooling/blob/master/LICENSE.md New BSD License
  */
 
@@ -13,6 +13,7 @@ use Zend\Expressive\Application;
 use Zend\Expressive\Middleware\ImplicitHeadMiddleware;
 use Zend\Expressive\Middleware\ImplicitOptionsMiddleware;
 use Zend\Expressive\Router\Route;
+use Zend\Stdlib\ConsoleHelper;
 use Zend\Stdlib\SplPriorityQueue;
 
 class Generator
@@ -30,7 +31,10 @@ class Generator
 
 use Zend\Expressive\Container\ErrorHandlerFactory;
 use Zend\Expressive\Container\ErrorResponseGeneratorFactory;
+use Zend\Expressive\Container\NotFoundDelegateFactory;
 use Zend\Expressive\Container\NotFoundHandlerFactory;
+use Zend\Expressive\Delegate\DefaultDelegate;
+use Zend\Expressive\Delegate\NotFoundDelegate;
 use Zend\Expressive\Middleware\ErrorResponseGenerator;
 use Zend\Expressive\Middleware\ImplicitHeadMiddleware;
 use Zend\Expressive\Middleware\ImplicitOptionsMiddleware;
@@ -40,6 +44,10 @@ use Zend\Stratigility\Middleware\OriginalMessages;
 
 return [
     'dependencies' => [
+        'aliases' => [
+            // Override the following to provide an alternate default delegate.
+            DefaultDelegate::class => NotFoundDelegate::class,
+        ],
         'invokables' => [
             ImplicitHeadMiddleware::class => ImplicitHeadMiddleware::class,
             ImplicitOptionsMiddleware::class => ImplicitOptionsMiddleware::class,
@@ -51,6 +59,8 @@ return [
             // Zend\Expressive\Container\WhoopsErrorResponseGeneratorFactory
             // in order to use Whoops for development error handling.
             ErrorResponseGenerator::class => ErrorResponseGeneratorFactory::class,
+            // Override the following to use an alternate "not found" delegate.
+            NotFoundDelegate::class => NotFoundDelegateFactory::class,
             NotFoundHandler::class => NotFoundHandlerFactory::class,
         ],
     ],
@@ -90,9 +100,9 @@ EOT;
 EOT;
 
     // @codingStandardsIgnoreStart
-    const TEMPLATE_PIPELINE_NO_PATH = '$app->%s(%s);';
+    const TEMPLATE_PIPELINE_NO_PATH = '$app->pipe(%s);';
 
-    const TEMPLATE_PIPELINE_WITH_PATH = '$app->%s(%s, %s);';
+    const TEMPLATE_PIPELINE_WITH_PATH = '$app->pipe(%s, %s);';
 
     const TEMPLATE_ROUTED_METHOD_NO_NAME = '$app->%s(\'%s\', %s)';
 
@@ -111,6 +121,19 @@ EOT;
      * @var string Root path against which paths are relative.
      */
     public $projectDir = '.';
+
+    /**
+     * @var ConsoleHelper
+     */
+    private $console;
+
+    /**
+     * @param ConsoleHelper $console
+     */
+    public function __construct(ConsoleHelper $console)
+    {
+        $this->console = $console;
+    }
 
     /**
      * @param string $configFile
@@ -202,12 +225,10 @@ EOT;
                 $pipeline[] = '$app->pipeRoutingMiddleware();';
                 $pipeline[] = sprintf(
                     self::TEMPLATE_PIPELINE_NO_PATH,
-                    'pipe',
                     $this->formatMiddleware(ImplicitHeadMiddleware::class)
                 );
                 $pipeline[] = sprintf(
                     self::TEMPLATE_PIPELINE_NO_PATH,
-                    'pipe',
                     $this->formatMiddleware(ImplicitOptionsMiddleware::class)
                 );
                 continue;
@@ -223,11 +244,18 @@ EOT;
             $path       = isset($spec['path']) ? (string) $spec['path'] : null;
             $middleware = $this->formatMiddleware($spec['middleware']);
             $error      = isset($spec['error']) ? (bool) $spec['error'] : false;
-            $method     = $error ? 'pipeErrorHandler' : 'pipe';
+
+            if ($error) {
+                $this->console->writeLine(sprintf(
+                    '<error>Encountered error middleware "%s"; did not add to pipeline</error>',
+                    $middleware
+                ), true, STDERR);
+                continue;
+            }
 
             $pipeline[] = (null === $path)
-                ? sprintf(self::TEMPLATE_PIPELINE_NO_PATH, $method, $middleware)
-                : sprintf(self::TEMPLATE_PIPELINE_WITH_PATH, $method, $this->createOptionValue($path), $middleware);
+                ? sprintf(self::TEMPLATE_PIPELINE_NO_PATH, $middleware)
+                : sprintf(self::TEMPLATE_PIPELINE_WITH_PATH, $this->createOptionValue($path), $middleware);
         }
 
         // Push the original messages middleware and error handler to the top