+++ test/GenerateProgrammaticPipelineFromConfig/GeneratorTest.php
@@ -10,8 +10,10 @@ namespace ZendTest\Expressive\Tooling\GenerateProgrammaticPipelineFromConfig;
 use org\bovigo\vfs\vfsStream;
 use org\bovigo\vfs\vfsStreamDirectory;
 use PHPUnit\Framework\TestCase;
+use Prophecy\Argument;
 use Zend\Expressive\Tooling\GenerateProgrammaticPipelineFromConfig\Generator;
 use Zend\Expressive\Tooling\GenerateProgrammaticPipelineFromConfig\GeneratorException;
+use Zend\Stdlib\ConsoleHelper;
 
 class GeneratorTest extends TestCase
 {
@@ -24,7 +26,8 @@ class GeneratorTest extends TestCase
     public function setUp()
     {
         $this->dir = vfsStream::setup('project');
-        $this->generator = new Generator();
+        $this->console = $this->prophesize(ConsoleHelper::class);
+        $this->generator = new Generator($this->console->reveal());
         $this->generator->projectDir = vfsStream::url('project');
     }
 
@@ -45,6 +48,14 @@ class GeneratorTest extends TestCase
     public function testGeneratesExpectedPipeline()
     {
         $this->generatePipeline();
+
+        $this->console
+            ->writeLine(
+                Argument::containingString('ErrorMiddleware'),
+                true,
+                STDERR
+            )
+            ->shouldBeCalled();
 
         $configFile = vfsStream::url('project/config/config.php');
         $this->generator->process($configFile);