+++ test/GenerateProgrammaticPipelineFromConfig/TestAsset/expected/config/autoload/programmatic-pipeline.global.php
@@ -5,7 +5,10 @@
 
 use Zend\Expressive\Container\ErrorHandlerFactory;
 use Zend\Expressive\Container\ErrorResponseGeneratorFactory;
+use Zend\Expressive\Container\NotFoundDelegateFactory;
 use Zend\Expressive\Container\NotFoundHandlerFactory;
+use Zend\Expressive\Delegate\DefaultDelegate;
+use Zend\Expressive\Delegate\NotFoundDelegate;
 use Zend\Expressive\Middleware\ErrorResponseGenerator;
 use Zend\Expressive\Middleware\ImplicitHeadMiddleware;
 use Zend\Expressive\Middleware\ImplicitOptionsMiddleware;
@@ -15,6 +18,10 @@ use Zend\Stratigility\Middleware\OriginalMessages;
 
 return [
     'dependencies' => [
+        'aliases' => [
+            // Override the following to provide an alternate default delegate.
+            DefaultDelegate::class => NotFoundDelegate::class,
+        ],
         'invokables' => [
             ImplicitHeadMiddleware::class => ImplicitHeadMiddleware::class,
             ImplicitOptionsMiddleware::class => ImplicitOptionsMiddleware::class,
@@ -26,6 +33,8 @@ return [
             // Zend\Expressive\Container\WhoopsErrorResponseGeneratorFactory
             // in order to use Whoops for development error handling.
             ErrorResponseGenerator::class => ErrorResponseGeneratorFactory::class,
+            // Override the following to use an alternate "not found" delegate.
+            NotFoundDelegate::class => NotFoundDelegateFactory::class,
             NotFoundHandler::class => NotFoundHandlerFactory::class,
         ],
     ],